refactor: move components out of the file
