+++ apis/nucleus/src/components/listbox/components/ListBoxRowColumn/components/ListBoxCheckbox.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { styled } from '@mui/material/styles';
 import { Checkbox } from '@mui/material';
-import CheckboxChecked from '../assets/CheckboxChecked';
+import CheckboxChecked from '../../../assets/CheckboxChecked';
 
 const PREFIX = 'ListBoxCheckbox';
 const borderRadius = 3;
