+++ apis/nucleus/src/components/listbox/components/ListBoxRowColumn/components/ValueField.jsx
@@ -1,0 +1,33 @@
+import React from 'react';
+import { Typography } from '@mui/material';
+import { joinClassNames } from '../helpers/operations';
+import classes from '../helpers/classes';
+import { excludedOrAlternative } from '../helpers/cell-states';
+
+export default function ValueField({
+  label,
+  color,
+  highlighted = false,
+  dense,
+  showGray = true,
+  checkboxes,
+  cell,
+  valueTextAlign,
+}) {
+  return (
+    <Typography
+      component="span"
+      variant="body2"
+      className={joinClassNames([
+        classes.labelText,
+        highlighted && classes.highlighted,
+        dense && classes.labelDense,
+        showGray && excludedOrAlternative({ cell, checkboxes }) && classes.excludedTextWithCheckbox,
+      ])}
+      color={color}
+      justifyContent={valueTextAlign}
+    >
+      <span style={{ whiteSpace: 'pre' }}>{label}</span>
+    </Typography>
+  );
+}
