+++ apis/nucleus/src/components/listbox/components/ListBoxRowColumn/index.jsx
@@ -1,0 +1,281 @@
+import React, { useState, useEffect, useCallback, useRef } from 'react';
+
+import { Grid } from '@mui/material';
+
+import Lock from '@nebula.js/ui/icons/lock';
+import Tick from '@nebula.js/ui/icons/tick';
+import getSegmentsFromRanges from '../listbox-highlight';
+import { getFieldKeyboardNavigation } from '../../interactions/listbox-keyboard-navigation';
+import classes from './helpers/classes';
+import { isAlternative, isExcluded } from './helpers/cell-states';
+import { joinClassNames } from './helpers/operations';
+import RowColRoot from './components/ListBoxRoot';
+import FieldWithRanges from './components/FieldWithRanges';
+import Field from './components/Field';
+import Histogram from './components/Histogram';
+import Frequency from './components/Frequency';
+import ItemGrid from './components/ItemGrid';
+
+function RowColumn({ index, rowIndex, columnIndex, style, data }) {
+  const {
+    onClick,
+    onMouseDown,
+    onMouseUp,
+    onMouseEnter,
+    pages,
+    isLocked,
+    column = false,
+    checkboxes = false,
+    textAlign,
+    direction,
+    layoutOptions = {},
+    freqIsAllowed,
+    isSingleSelect,
+    actions,
+    frequencyMax = '',
+    histogram = false,
+    keyboard,
+    showGray = true,
+    columnCount = 1,
+    rowCount = 1,
+    dataOffset,
+    focusListItems,
+    listCount,
+    itemPadding,
+  } = data;
+
+  const { dense = false, dataLayout = 'singleColumn', layoutOrder } = layoutOptions;
+
+  let cellIndex;
+  let styles;
+  const count = { max: null, currentIndex: null };
+
+  if (typeof rowIndex === 'number' && typeof columnIndex === 'number') {
+    if (layoutOrder === 'row') {
+      cellIndex = rowIndex * columnCount + columnIndex;
+      count.max = rowCount;
+      count.currentIndex = rowIndex;
+    } else {
+      cellIndex = columnIndex * rowCount + rowIndex;
+      count.max = columnCount;
+      count.currentIndex = columnIndex;
+    }
+    const padding = 0;
+    styles = {
+      ...style,
+      left: padding + (columnIndex === 0 ? style.left : Number(style.left) + columnIndex * padding),
+      // right: columnIndex === columnCount ? style.right : Number(style.right) + columnIndex * padding,
+      top: rowIndex === 0 ? style.top : Number(style.top) + rowIndex * padding,
+    };
+  } else {
+    cellIndex = index;
+    count.max = listCount;
+    count.currentIndex = index;
+    styles = { ...style };
+  }
+  cellIndex += dataOffset;
+
+  const rowRef = useRef(null);
+  useEffect(() => {
+    if (rowRef.current !== null) {
+      if (count.currentIndex === 0 && focusListItems.first) {
+        rowRef.current.focus();
+        focusListItems.setFirst(false);
+      }
+      if (count.currentIndex === count.max - 1 && focusListItems.last) {
+        rowRef.current.focus();
+        focusListItems.setLast(false);
+      }
+    }
+  }, [rowRef.current]);
+
+  const handleKeyDownCallback = useCallback(getFieldKeyboardNavigation({ ...actions, focusListItems }), [actions]);
+
+  const [isSelected, setSelected] = useState(false);
+  const [cell, setCell] = useState();
+
+  const [classArr, setClassArr] = useState([]);
+
+  useEffect(() => {
+    if (!pages) {
+      return;
+    }
+    let c;
+    const page = pages.filter((p) => p.qArea.qTop <= cellIndex && cellIndex < p.qArea.qTop + p.qArea.qHeight)[0];
+    if (page) {
+      const area = page.qArea;
+      if (cellIndex >= area.qTop && cellIndex < area.qTop + area.qHeight) {
+        [c] = page.qMatrix[cellIndex - area.qTop];
+      }
+    }
+    setCell(c);
+  }, [pages]);
+
+  useEffect(() => {
+    if (!cell) {
+      return;
+    }
+    const selected = cell.qState === 'S' || cell.qState === 'XS' || cell.qState === 'L';
+    setSelected(selected);
+
+    const clazzArr = [column ? classes.column : classes.row];
+    if (!(histogram && (dense || checkboxes))) clazzArr.push(classes.rowBorderBottom);
+    if (!checkboxes) {
+      if (cell.qState === 'XS') {
+        clazzArr.push(showGray ? classes.XS : classes.S);
+      } else if (cell.qState === 'S' || cell.qState === 'L') {
+        clazzArr.push(classes.S);
+      } else if (showGray && isAlternative(cell)) {
+        clazzArr.push(classes.A);
+      } else if (showGray && isExcluded(cell)) {
+        clazzArr.push(classes.X);
+      }
+    }
+    setClassArr(clazzArr);
+  }, [cell && cell.qState, histogram, dense, checkboxes]);
+
+  if (!cell) {
+    return null; // prevent rendering empty rows
+  }
+
+  const isNumeric = !['NaN', undefined].includes(cell?.qNum);
+  let valueTextAlign;
+  const isAutoTextAlign = !textAlign || textAlign.auto;
+  const dirToTextAlignMap = {
+    rtl: 'right',
+    ltr: 'left',
+  };
+
+  if (isAutoTextAlign) {
+    valueTextAlign = isNumeric ? 'right' : dirToTextAlignMap[direction];
+  } else {
+    valueTextAlign = textAlign?.align || 'left';
+  }
+
+  const preventContextMenu = (event) => {
+    if (checkboxes) {
+      // Event will not propagate in the checkbox/radiobutton case
+      onClick(event);
+    }
+    event.preventDefault();
+  };
+
+  const isGridCol = dataLayout === 'grid' && layoutOrder === 'column';
+
+  const label = cell?.qText ?? '';
+
+  // Search highlights. Split up labelText span into several and add the highlighted class to matching sub-strings.
+  const ranges =
+    (cell && cell.qHighlightRanges && cell.qHighlightRanges.qRanges.sort((a, b) => a.qCharPos - b.qCharPos)) || [];
+
+  const labels = getSegmentsFromRanges(label, ranges);
+
+  const iconStyles = {
+    alignItems: 'center',
+    display: 'flex',
+    fontSize: '8px',
+  };
+
+  const showLock = isSelected && isLocked;
+  const showTick = !checkboxes && isSelected && !isLocked;
+
+  const cellStyle = {
+    display: 'flex',
+    alignItems: 'center',
+    minWidth: 0,
+    flexGrow: 1,
+    padding: checkboxes ? 0 : undefined,
+    justifyContent: valueTextAlign,
+  };
+
+  const isFirstElement = index === 0;
+  const flexBasisVal = checkboxes ? 'auto' : 'max-content';
+
+  return (
+    <RowColRoot
+      className={classes.barContainer}
+      flexBasisProp={flexBasisVal}
+      style={styles}
+      isGridCol={isGridCol}
+      isGridMode={dataLayout === 'grid'}
+      dense={dense}
+    >
+      <ItemGrid
+        ref={rowRef}
+        container
+        dataLayout={dataLayout}
+        layoutOrder={layoutOrder}
+        itemPadding={itemPadding}
+        gap={0}
+        className={joinClassNames(['value', ...classArr])}
+        classes={{
+          root: classes.fieldRoot,
+        }}
+        onClick={onClick}
+        onMouseDown={onMouseDown}
+        onMouseUp={onMouseUp}
+        onMouseEnter={onMouseEnter}
+        onKeyDown={handleKeyDownCallback}
+        onContextMenu={preventContextMenu}
+        role={column ? 'column' : 'row'}
+        tabIndex={isFirstElement && (!keyboard.enabled || keyboard.active) ? 0 : -1}
+        data-n={cell && cell.qElemNumber}
+      >
+        <Histogram
+          cell={cell}
+          histogram={histogram}
+          checkboxes={checkboxes}
+          isSelected={isSelected}
+          frequencyMax={frequencyMax}
+        />
+        <Grid
+          item
+          style={cellStyle}
+          className={joinClassNames([classes.cell, classes.selectedCell])}
+          title={`${label}`}
+        >
+          {ranges.length === 0 ? (
+            <Field
+              label={label}
+              color="inherit"
+              qElemNumber={cell.qElemNumber}
+              isSelected={isSelected}
+              dense={dense}
+              cell={cell}
+              isGridCol={isGridCol}
+              showGray={showGray}
+              isSingleSelect={isSingleSelect}
+              checkboxes={checkboxes}
+            />
+          ) : (
+            <FieldWithRanges
+              labels={labels}
+              checkboxes={checkboxes}
+              dense={dense}
+              showGray={showGray}
+              color="inherit" // TODO: Check this
+              qElemNumber={cell.qElemNumber}
+              isSelected={isSelected}
+              cell={cell}
+              isGridCol={isGridCol}
+              isSingleSelect={isSingleSelect}
+              valueTextAlign={valueTextAlign}
+            />
+          )}
+        </Grid>
+
+        {freqIsAllowed && <Frequency cell={cell} checkboxes={checkboxes} dense={dense} showGray={showGray} />}
+
+        {!checkboxes && (
+          <Grid item className={classes.icon}>
+            {!showLock && !showTick && <span style={{ minWidth: '12px' }} />}
+            {showLock && <Lock style={iconStyles} size="small" />}
+            {showTick && <Tick style={iconStyles} size="small" />}
+          </Grid>
+        )}
+      </ItemGrid>
+    </RowColRoot>
+  );
+}
+
+export default RowColumn;
